+++ Assets/Scripts/EntitySpawner.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.Events;
 
-public class TakeDamageEvent : UnityEvent<int>
+public class TakeDamageEvent : UnityEvent<long>
 {
 
 }
