+++ Assets/Scripts/RockUpgrade.cs
@@ -9,13 +9,13 @@ public class RockUpgrade : Upgrade
         stats.rockDamage += GetDamage();
     }
 
-    private int GetDamage() => lvl > 0 ? Mathf.RoundToInt(50000f * Mathf.Pow(1.07f, lvl)) : 0;
+    private long GetDamage() => lvl > 0 ? Mathf.RoundToInt(50000f * Mathf.Pow(1.07f, lvl)) : 0;
 
     public override string GetDescription() => $"Shoot rock with {PlayerInfo.GetAdaptedInt(GetDamage())} dmg every 3 seconds";
 
     public override string GetName() => "Rock";
 
-    public override int GetUpgradeCost() => Mathf.RoundToInt(500000f * Mathf.Pow(1.07f, lvl));
+    public override long GetUpgradeCost() => Mathf.RoundToInt(500000f * Mathf.Pow(1.07f, lvl));
     //public override int GetUpgradeCost() => 1;
